+++ addons/yat/src/commands/builtin/Set.cs
@@ -1,5 +1,5 @@
 using YAT.Helpers;
-using YAT.Overlay.Components;
+using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Commands
 {