+++ addons/yat/src/helpers/LogHelper.cs
@@ -1,5 +1,5 @@
 using Godot;
-using YAT.Overlay.Components;
+using YAT.Overlay.Components.Terminal;
 
 namespace YAT.Helpers
 {