+++ example/SetCube.cs
@@ -1,6 +1,6 @@
 using Godot;
 using YAT.Commands;
-using YAT.Overlay.Components;
+using YAT.Overlay.Components.Terminal;
 
 [Extension("cubecolor", "Sets the cube's color.", "[b]Usage[/b]: setcube [i]color[/i]")]
 public partial class SetCube : IExtension